% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDrugCellInteractions.R
\name{analyzeDrugCellInteractions}
\alias{analyzeDrugCellInteractions}
\title{Analyze Drug-Cell Interactions and calculate Tri-score}
\usage{
analyzeDrugCellInteractions(
  ALMANAC_sub_tissue_converted,
  tissue_folder,
  unique_cellnames,
  unique_drugnames
)
}
\arguments{
\item{ALMANAC_sub_tissue_converted}{Data frame containing drug and cell line data.}

\item{tissue_folder}{Path to the folder for saving the output files.}
}
\value{
NULL Invisible; this function saves several files to disk.
}
\description{
This function calculates mean resistance indices, classifies synergy and antagonism,
and classifies cell line sensitivity to drug combinations, generate Tri-score.
}
