% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDysregulatedPathways.R
\name{analyzeDysregulatedPathways}
\alias{analyzeDysregulatedPathways}
\title{Analyze Dysregulated Pathways from GEO Data}
\usage{
analyzeDysregulatedPathways(
  geoID,
  sensitivity_classification,
  pathway_gene_set,
  unique_drugnames,
  tissue_folder,
  tissue,
  cancer
)
}
\arguments{
\item{geoID}{GEO Series ID to fetch data for analysis.}

\item{tissue_folder}{Directory path to store intermediate and final results.}

\item{cancer}{Specific cancer type for analysis.}

\item{gene_sets}{List of gene sets for pathway analysis.}
}
\value{
Returns a list of dysregulated pathways for further analysis.
}
\description{
This function retrieves and analyzes gene expression data from GEO to identify dysregulated pathways based on differential expression analysis.
}
