% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_AUC_prediction_result.R
\name{calculate_performance_metrics}
\alias{calculate_performance_metrics}
\title{Calculate AUROC and AUPRC}
\usage{
calculate_performance_metrics(tissue_folder, tissue, threshold = 0.99)
}
\arguments{
\item{tissue_folder}{The folder containing the RDS files.}

\item{tissue}{The specific tissue type for which calculations are performed.}
}
\value{
A list containing AUROC and AUPRC values.
}
\description{
This function calculates the Area Under the Receiver Operating Characteristic (AUROC)
and the Area Under the Precision-Recall Curve (AUPRC) for a given set of drug combinations
and their associated probabilities. The function also outputs prediction results.
}
