% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanRI.R
\name{calculateMeanRI}
\alias{calculateMeanRI}
\title{Calculate Mean Resistance Index}
\usage{
calculateMeanRI(data, cellnames, drugnames)
}
\arguments{
\item{data}{Data frame containing drug and cell line data.}

\item{cellnames}{Vector of unique cell line names.}

\item{drugnames}{Vector of unique drug names.}
}
\value{
Data frame with mean resistance indices.
}
\description{
Calculate Mean Resistance Index
}
