% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTriScores.R
\name{calculateTriScores}
\alias{calculateTriScores}
\title{Calculate Tri-Scores for Drug Combinations}
\usage{
calculateTriScores(dataset, synergy_list, thresholds)
}
\arguments{
\item{dataset}{dataset frame with drug interaction dataset.}

\item{synergy_list}{List of dataset frames containing synergy summaries for different thresholds.}

\item{thresholds}{Numeric vector of thresholds for scoring.}
}
\value{
dataset frame with tri-scores for each drug combination.
}
\description{
Calculate Tri-Scores for Drug Combinations
}
