% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySensitivity.R
\name{classifySensitivity}
\alias{classifySensitivity}
\title{Classify Sensitivity Based on Resistance Indices}
\usage{
classifySensitivity(ri_data, tissue_folder)
}
\arguments{
\item{ri_data}{Data frame containing resistance indices.}

\item{tissue_folder}{Folder path for saving results.}
}
\value{
Data frame with sensitivity classifications.
}
\description{
Classify Sensitivity Based on Resistance Indices
}
