% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseProbes.R
\name{collapseProbes}
\alias{collapseProbes}
\title{Collapse probes to genes}
\usage{
collapseProbes(
  x,
  method = "variance",
  topTable = NULL,
  colNameOfStat = NULL,
  colNameOfGeneSymbol = "SYMBOL"
)
}
\arguments{
\item{x}{ExpressionSet needs to reduced probes / genes.}
}
\value{
Reduced ExpressionSet.
}
\description{
This function summarize several probe measurements corresponding to a single gene.
}
