% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseProbesAndValidate.R
\name{collapseProbesAndValidate}
\alias{collapseProbesAndValidate}
\title{Collapse Probes and Validate Data}
\usage{
collapseProbesAndValidate(eset, method = "variance")
}
\arguments{
\item{eset}{An ExpressionSet object to be processed.}

\item{method}{A character string indicating the collapsing method, defaults to 'variance'.}
}
\value{
A modified ExpressionSet with collapsed probes.
}
\description{
Collapses probes in an ExpressionSet using a specified method and ensures gene symbols are correctly assigned.
}
