% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalizeResults.R
\name{finalizeResults}
\alias{finalizeResults}
\title{Convert list of interesting_pathways to data frame and save it as csv}
\usage{
finalizeResults(pathway_results, pathway_gene_set, tissue_folder, tissue)
}
\arguments{
\item{pathway_results}{A list of data frames where each data frame corresponds to a different drug, containing columns 'pathway' and 'NES'.}

\item{pathway_gene_set}{A named list of pathway gene sets.}

\item{tissue_folder}{A string specifying the folder path for saving the output CSV file.}

\item{tissue}{A string specifying the tissue type, used in naming the output file.}
}
\value{
A data frame summarizing the pathway analysis results for each drug, also saved as a CSV file in the specified location.
}
\description{
This function processes pathway results for different drugs and generates a CSV file summarizing the results.
}
