% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraph.R
\name{generateGraph}
\alias{generateGraph}
\title{Generate and Analyze Graph from Drug Pathway Data with igraph package}
\usage{
generateGraph(drug_pathway_df, tissue_folder, tissue)
}
\arguments{
\item{drug_pathway_df}{A data frame where rows correspond to probes and columns to different drugs.}

\item{tissue_folder}{A string specifying the directory to save output files.}

\item{tissue}{A string specifying the tissue type, used in naming the output files.}
}
\value{
Generates a graph object and saves it as an RDS file
}
\description{
This function prepares a graph from drug pathway data.
}
