% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_preprocess_data.R
\name{load_and_preprocess_data}
\alias{load_and_preprocess_data}
\title{Load and Preprocess Drug and Cell Line Data}
\usage{
load_and_preprocess_data(
  drug_data_path = NULL,
  cell_data_path = NULL,
  summary_df_ALMANAC_path = NULL,
  summary_df_api_path = NULL,
  pathway_gene_set_path = NULL,
  tissue_cancer_infosheet_path = NULL,
  env = .GlobalEnv
)
}
\arguments{
\item{drug_data_path}{Path to the drug data CSV file from DrugComb.}

\item{cell_data_path}{Path to the cell line data CSV file from DrugComb.}

\item{summary_df_ALMANAC_path}{Path to the summary data filtered by ALMANAC, CSV file from DrugComb.}

\item{summary_df_api_path}{Path to the summary data download by api, CSV file from DrugComb.}

\item{pathway_gene_set_path}{Path to the pathway gene set file, download from https://www.gsea-msigdb.org/.}

\item{tissue_cancer_infosheet_path}{Path to the cancer-tissue info sheet of GSE entry, CSV file.}

\item{env}{Environment in which to assign the dataframes (default is the global environment).}
}
\value{
NULL Invisible; this function release variables to .GlobalEnv.
}
\description{
This function takes paths to drug and cell line data files and optionally other data files,
loads them, applies necessary transformations, and assigns each dataframe to a variable
in the specified environment.
}
