% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeGroupData.R
\name{normalizeGroupData}
\alias{normalizeGroupData}
\title{Normalize Group Data}
\usage{
normalizeGroupData(eset_final, pdata, cancer)
}
\arguments{
\item{pdata}{A DataFrame of phenotypic data from an ExpressionSet.}

\item{cancer}{A character string specifying the cancer type to filter the data by.}
}
\value{
A DataFrame with normalized group information.
}
\description{
Extracts and normalizes experimental conditions from the phenotypic data of an ExpressionSet.
}
