% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_2columns.R
\name{order_2columns}
\alias{order_2columns}
\title{Reorder Columns If Necessary}
\usage{
order_2columns(x)
}
\arguments{
\item{x}{A dataframe with at least two columns.}
}
\value{
A dataframe with possibly reordered columns.
}
\description{
This function reorders the first two columns of a dataframe if the first column
is greater than the second.
}
