% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performPathwayAnalysis.R
\name{performPathwayAnalysis}
\alias{performPathwayAnalysis}
\title{Perform Pathway Analysis}
\usage{
performPathwayAnalysis(
  group_info,
  eset_sub_tissue,
  tissue_folder,
  env = environment(),
  sensitivity_classification,
  pathway_gene_set,
  unique_drugnames
)
}
\arguments{
\item{group_info}{A DataFrame containing normalized group information for constructing model matrices.}

\item{eset_sub_tissue}{An ExpressionSet to be used for differential expression analysis.}

\item{tissue_folder}{A string indicating the directory where results should be saved.}

\item{gene_sets}{A list of gene sets for performing pathway analysis.}
}
\value{
A list containing the results of the pathway analysis for each drug.
}
\description{
Executes differential expression analysis using 'limma' and pathway analysis using 'fgsea'.
It uses parallel processing to efficiently handle computations.
}
