% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_by_cluster.R
\name{prediction_by_cluster}
\alias{prediction_by_cluster}
\title{Predict drug synergy from cluster}
\usage{
prediction_by_cluster(
  tissue_folder,
  threshold_graph = 95,
  tri_scores,
  graph_pathway
)
}
\arguments{
\item{tissue_folder}{The folder containing the RDS files.}

\item{threshold_graph}{The threshold to use for the initial filter on graph edges.}
}
\value{
A list of drug combinations with their synergy probabilities.
}
\description{
This function calculates synergy scores for drug combinations
based on clustering of in graph.
It outputs a list of probabilities indicating potential synergies.
}
