% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTissueCancerInfo.R
\name{processTissueCancerInfo}
\alias{processTissueCancerInfo}
\title{Process Tissue Cancer Information and Convert Drug and Cell IDs}
\usage{
processTissueCancerInfo(
  tissue_cancer_infosheet,
  summary_df_ALMANAC,
  summary_df_api,
  DC_drug,
  DC_cell,
  row_number
)
}
\arguments{
\item{tissue_cancer_infosheet}{A data frame containing tissue and cancer information.}

\item{summary_df_ALMANAC}{A data frame summarizing ALMANAC data by tissue.}

\item{summary_df_api}{A data frame from API summarizing additional details.}

\item{DC_drug}{A data frame containing drug information including 'id' and 'stitch_name'.}

\item{DC_cell}{A data frame containing cell line information including 'id' and 'name'.}

\item{row_number}{The row number of cancer from tissue_cancer_infosheet to process.}
}
\value{
A data frame of the processed ALMANAC tissue data with converted IDs.
}
\description{
This function processes tissue cancer information from specified rows of a data frame,
creates necessary directories, subsets and cleans the data, and converts drug and cell IDs.
}
