% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveGEOData.R
\name{retrieveGEOData}
\alias{retrieveGEOData}
\title{Retrieve GEO Data}
\usage{
retrieveGEOData(geoID, tissue_folder)
}
\arguments{
\item{geoID}{A character string representing the GEO Series ID.}

\item{tissue_folder}{A character string indicating the directory path for saving the retrieved data.}
}
\value{
An ExpressionSet containing the GEO data.
}
\description{
Fetches and caches GEO datasets based on the provided GEO Series ID, ensuring data is only downloaded once.
}
